# Fused QKV bitnet