--- a/mlx_lm/models/bitlinear_layers.py
+++ b/mlx_lm/models/bitlinear_layers.py
@@ -96,7 +96,7 @@ class BitLinear(nn.Module):
             inputs=[x_flattened.astype(self.dtype), packed_weights, self.weight_scale, self.invert_weight_scales],
             template=[("batch_size", total_batch_elements), ("in_features", in_features), ("out_features", out_features)],
             grid=(total_batch_elements * out_features, 1, 1),
-            threadgroup=(min(32, total_batch_elements * out_features), 1, 1), # SIMD width is 32 threads
+            threadgroup=(32, 1, 1), # SIMD width is 32 threads
             output_shapes=[(total_batch_elements, out_features)],
             output_dtypes=[self.dtype],
         )