--- a/mlx_lm/models/bitlinear_layers.py
+++ b/mlx_lm/models/bitlinear_layers.py
@@ -1,5 +1,6 @@
 import mlx.core as mx
 import mlx.nn as nn
+from mlx.nn.layers.quantized import QuantizedLinear
 
 
 class BitLinear(nn.Module):
@@ -139,5 +140,28 @@ class BitLinear(nn.Module):
         # Add bias if present
         if self.bias is not None:
             y = mx.add(y, self.bias)
-
         return y.astype(org_dtype)
+
+
+class QuantAndBitLinear(nn.Linear):
+    """
+    A Linear layer that can be converted to a quantized and bitlinear version.
+    """
+
+    def to_quantized(self, method: str = None, group_size: int = 64, bits: int = 4, **kwargs):
+
+        if method is None or group_size is None or bits is None:
+            return QuantizedLinear.from_linear(self, group_size, bits)
+
+        if method == "bitnet":
+            bitlinear = BitLinear(
+                in_features=self.weight.shape[1],
+                out_features=self.weight.shape[0],
+                bias= getattr(self, "bias", None) is not None,
+                invert_weight_scales=True,
+                **kwargs
+            )
+            return bitlinear
+        else:
+            raise ValueError(f"Unknown quantization method: {method}")
+