![image](https://github.com/user-attachments/assets/4d5b685d-3187-43ca-a65a-40f52fbc4381)
This PR introduces a fused QKV (Query-Key-Value) implementation in the attention module for BitNet-1.58-2B on MLX. The fusion significantly improves prompt and generation speed by around 10%.

Key Changes:

- Added support for fused QKV projection in the attention layer.
- Updated model forward pass to conditionally use the fused path when enabled.

Benchmarked performance improvements (M3 Max)"
- Prompt Processing: ↑ from 128.77 to 139.94 tokens/sec
- Generation Speed: ↑ from 67.05 to 73.12 tokens/sec
- MLX Fused QKV vs BitNet 4T = 27.6% faster generation, 137% faster prompt